rename: 컴포넌트 내 screens 폴더 생성 후 폴더 구조 변경
