+++ src/screens/Tab/TransactionScreen.tsx
@@ -3,10 +3,10 @@ import {View, StyleSheet} from 'react-native';
 import 'dayjs/locale/ko';
 import {theme} from 'styles';
 import ScreenContainer from 'components/@common/ScreenContainer';
-import MonthHeader from 'components/TransactionScreen/MonthHeader';
-import FinancialOverview from 'components/TransactionScreen/FinancialOverview';
-import FinancialCalendar from 'components/TransactionScreen/FinancialCalendar';
-import TransactionList from 'components/TransactionScreen/TransactionList';
+import MonthHeader from 'components/screens/TransactionScreen/MonthHeader';
+import FinancialOverview from 'components/screens/TransactionScreen/FinancialOverview';
+import FinancialCalendar from 'components/screens/TransactionScreen/FinancialCalendar';
+import TransactionList from 'components/screens/TransactionScreen/TransactionList';
 
 export default function TransactionScreen() {
   const [currentDate, setCurrentDate] = useState(new Date());
